+++ config/routes.rb
@@ -223,6 +223,7 @@ Rails.application.routes.draw do
         end
         resources :referee_qualification_types, only: %i[index create update destroy]
         resources :referee_license_levels, only: %i[index create update destroy]
+        resources :penalty_codes, only: %i[index create update destroy]
         resources :referee_course_imports, only: %i[index show create destroy] do
           post :submit, on: :member
         end